worker/storageprovisioner: reschedule destroy
If DestroyVolumes returns a volume-specific error,
then we reschedule destruction of that volume for
later and report the failure through status.

+++ worker/storageprovisioner/common.go
@@ -72,6 +72,9 @@ func attachmentLife(ctx *context, ids []params.MachineStorageId) (
 
 // removeEntities removes each specified Dead entity from state.
 func removeEntities(ctx *context, tags []names.Tag) error {
+	if len(tags) == 0 {
+		return nil
+	}
 	logger.Debugf("removing entities: %v", tags)
 	errorResults, err := ctx.life.Remove(tags)
 	if err != nil {
@@ -100,6 +103,16 @@ func removeAttachments(ctx *context, ids []params.MachineStorageId) error {
 		}
 	}
 	return nil
+}
+
+// setStatus sets the given entity statuses, if any. If setting
+// the status fails the error is logged but otherwise ignored.
+func setStatus(ctx *context, statuses []params.EntityStatus) {
+	if len(statuses) > 0 {
+		if err := ctx.statusSetter.SetStatus(statuses); err != nil {
+			logger.Errorf("failed to set status: %v", err)
+		}
+	}
 }
 
 var errNonDynamic = errors.New("non-dynamic storage provider")
